services/document-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sentence-transformers==2.5.1
weaviate-client==3.26.0
pypdf==4.0.2
python-multipart==0.0.9
requests==2.31.0
python-docx==1.1.0
openpyxl==3.1.2
